+++ lib/modal_route/transitions/build_transitions/widgets/slide.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
 
 /// Positions
 ///
@@ -19,7 +19,7 @@ enum Positions {
   // Offset(0, -1 / 3),
   up(
     primaryOffset: Offset(0, 1),
-    secondaryOffset: Offset(0, 0),
+    secondaryOffset: Offset.zero,
   ),
   down(
     primaryOffset: Offset(0, -1),