+++ lib/route_observer.dart
@@ -92,9 +92,21 @@ class _RouteObserverProviderState
   void didChangeDependencies() {
     super.didChangeDependencies();
 
-    final ModalRoute? modalRoute = ModalRoute.of(context);
-    if (modalRoute is ModalRoute) {
+    final modalRoute = ModalRoute.of<void>(context);
+    if (modalRoute != null) {
       widget.routeObserver.subscribe(this, modalRoute);
+    }
+  }
+
+  @override
+  void didUpdateWidget(RouteObserverProvider oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.routeObserver != oldWidget.routeObserver) {
+      oldWidget.routeObserver.unsubscribe(this);
+      final modalRoute = ModalRoute.of<void>(context);
+      if (modalRoute != null) {
+        widget.routeObserver.subscribe(this, modalRoute);
+      }
     }
   }