+++ lib/sheet/modal_base_params.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/widgets.dart';
+
 base class ModalBaseParams {
   const ModalBaseParams();
 }
@@ -53,7 +55,7 @@ base class ModalDraggableScrollableSheetParams extends ModalBaseParams {
   /// The initial fractional value of the parent container's height to use when
   /// displaying the widget.
   ///
-  /// Rebuilding the sheet with a new [initialChildSize] will only move
+  /// Rebuilding the sheet with a new `initialChildSize` will only move
   /// the sheet to the new value if the sheet has not yet been dragged since it
   /// was first built or since the last call to [DraggableScrollableActuator.reset].
   ///
@@ -95,26 +97,26 @@ base class ModalDraggableScrollableSheetParams extends ModalBaseParams {
   ///
   /// Rebuilding the sheet with snap newly enabled will immediately trigger a
   /// snap unless the sheet has not yet been dragged away from
-  /// [initialChildSize] since first being built or since the last call to
+  /// `initialChildSize` since first being built or since the last call to
   /// [DraggableScrollableActuator.reset].
   final bool snap;
 
   /// A list of target sizes that the widget should snap to.
   ///
   /// Snap sizes are fractional values of the parent container's height. They
-  /// must be listed in increasing order and be between [minChildSize] and
-  /// [maxChildSize].
+  /// must be listed in increasing order and be between `minChildSize` and
+  /// `maxChildSize`.
   ///
-  /// The [minChildSize] and [maxChildSize] are implicitly included in snap
+  /// The `minChildSize` and `maxChildSize` are implicitly included in snap
   /// sizes and do not need to be specified here. For example, `snapSizes = [.5]`
-  /// will result in a sheet that snaps between [minChildSize], `.5`, and
-  /// [maxChildSize].
+  /// will result in a sheet that snaps between `minChildSize`, `.5`, and
+  /// `maxChildSize`.
   ///
   /// Any modifications to the [snapSizes] list will not take effect until the
   /// `build` function containing this widget is run again.
   ///
   /// Rebuilding with a modified or new list will trigger a snap unless the
-  /// sheet has not yet been dragged away from [initialChildSize] since first
+  /// sheet has not yet been dragged away from `initialChildSize` since first
   /// being built or since the last call to [DraggableScrollableActuator.reset].
   final List<double>? snapSizes;