+++ lib/sheet/sheet.dart
@@ -1,6 +1,6 @@
 library;
 
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 import '../browser.dart';